[Skia] Implement dashed lines support https://bugs.webkit.org/show_bug.cgi?id=269471
Reviewed by Adrian Perez de Castro and Don Olmstead.

Store the dash data (the array, and the offset) into an auxiliary struct
field. If populated, use that to create a SkDashPathEffect, and apply
the effect to the SkPaint used for stroke.

The constructor of SkDashPathEffect has the exact same semantics that
of the array and offset passed to GraphicsContextSkia::setLineDash(),
which makes it extra easy since no conversion needs to happen.

* Source/WebCore/platform/graphics/skia/GraphicsContextSkia.cpp:
(WebCore::GraphicsContextSkia::createStrokePaint const):
(WebCore::GraphicsContextSkia::setLineDash):
* Source/WebCore/platform/graphics/skia/GraphicsContextSkia.h:

Canonical link: https://commits.webkit.org/274758@main