[cocoa] Add a preference for using the system's default network loader https://bugs.webkit.org/show_bug.cgi?id=269542 rdar://123064383
Reviewed by NOBODY (OOPS!).

Add an additional preference that controls whether WebKit prefers using the
system's default network loader. This complements the existing preferences for
enabling the new experimental loader.

If CFNetworkNetworkLoaderEnabled is true, then the experimental network loader
is enabled; else, if that preference is not true, and
CFNetworkDefaultNetworkLoaderEnabled is not false, then we use the system
default. Otherwise, if both CFNetworkNetworkLoaderEnabled and
CFNetworkDefaultNetworkLoaderEnabled are false, then we disable using the
experimental loader.

Tested manually.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebKit/NetworkProcess/cocoa/NetworkSessionCocoa.mm:
(WebKit::NetworkSessionCocoa::NetworkSessionCocoa):
* Source/WebKit/UIProcess/WebsiteData/Cocoa/WebsiteDataStoreCocoa.mm:
(WebKit::WebsiteDataStore::useNetworkLoader):
* Tools/WebKitTestRunner/TestController.cpp:
(WTR::TestController::resetPreferencesToConsistentValues):