[git-webkit revert] Should error when radar is not found instead of creating bug with radar title https://bugs.webkit.org/show_bug.cgi?id=268934 rdar://problem/122490424
Reviewed by Jonathan Bedard.

Returns an error when an issue is not found when a link is passed in.
Previous behaviour meant that the link would be used as a title to construct a new bug.

* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/revert.py:
(Revert):
(Revert.get_issue_info): Adds check so that radar and bugzilla links are not used as titles.
(Revert.create_revert_commit_msg): Consolidates issue logic.

Canonical link: https://commits.webkit.org/274885@main