[visionOS] Attach a FigVideoTarget to AVPlayers when a receiver endpoint is available https://bugs.webkit.org/show_bug.cgi?id=269561 rdar://123073265
Reviewed by Jer Noble.

When MediaPlayer has a VideoReceiverEndpoint, use it to create a FigVideoTarget and add it to
MediaPlayerPrivateAVFoundationObjC's AVPlayer. In order for LinearMediaPlayer to emit a
VideoReceiverEndpoint, also created an Entity with a caption layer rather than an AVPlayer.

* Source/WebCore/PAL/pal/cocoa/MediaToolboxSoftLink.cpp:
* Source/WebCore/PAL/pal/cocoa/MediaToolboxSoftLink.h:
* Source/WebCore/PAL/pal/spi/cocoa/MediaToolboxSPI.h:
Declared and soft-linked FigVideoTargetCreateWithVideoReceiverEndpointID.

* Source/WebCore/platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.h:
* Source/WebCore/platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
When non-null, used the VideoReceiverEndpoint to create a FigVideoTargetRef, added it to AVPlayer,
and removed the player from the AVPlayerLayer. When null, removed the video target from the player
and re-attached the player to the AVPlayerLayer.

* Source/WebKit/Platform/ios/PlaybackSessionInterfaceLMK.mm:
Cleared the VideoReceiverEndpoint when exiting fullscreen. Drive-by fix: set m_playerDelegate as
m_player's delegate.

* Source/WebKit/Platform/ios/VideoPresentationInterfaceLMK.mm:
Set m_player's captionLayer prior to creating a view controller.

* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.sb.in:
Allowed lookup of mach port com.apple.coremedia.videotarget.xpc.

* Source/WebKit/WebKitSwift/LinearMediaKit/LinearMediaPlayer.swift:
Stopped publishing captionLayer and used it to create an Entity instead.

Canonical link: https://commits.webkit.org/274909@main