[Skia] Implement outset drop shadows https://bugs.webkit.org/show_bug.cgi?id=269596
Reviewed by Nikolas Zimmermann.

Whenever the GraphicsContext has a drop shadow set, use that to create
and set a SkImageFilters::DropShadow() image effect to the fill paint.
Stroke paints are unaffected.

The SkImageFilters::DropShadow() receives a blur sigma instead of radius
and the formal definition of that [1] states that sigma is half of the
blur radius, so convert between these values.

This only implements outset shadows. Inset shadows won't be rendered
anyway because GraphicsContextSkia::fillRectWithRoundedHole() is not
implemented yet.

[1] https://html.spec.whatwg.org/multipage/canvas.html#shadows

* Source/WebCore/platform/graphics/skia/GraphicsContextSkia.cpp:
(WebCore::GraphicsContextSkia::createFillPaint const):

Canonical link: https://commits.webkit.org/274928@main