[macOS] Screen or Window sharing in Safari stops working once paused, resume option does not work - WebRTC getDisplayMedia https://bugs.webkit.org/show_bug.cgi?id=269566 rdar://122187886
Reviewed by Eric Carlson.

When muting the source, we stop it.
At that point, we were forgetting which content filter we were using, which is necessary to restart the stream.
We are now storing it in ScreenCaptureKitCaptureSource by querying ScreenCaptureKitSharingSessionManager.
When unmuting, we already have it and we can construct successfully a new session source.

* Source/WebCore/platform/mediastream/mac/ScreenCaptureKitCaptureSource.h:
* Source/WebCore/platform/mediastream/mac/ScreenCaptureKitCaptureSource.mm:
(WebCore::ScreenCaptureKitCaptureSource::startContentStream):
* Source/WebCore/platform/mediastream/mac/ScreenCaptureKitSharingSessionManager.h:
* Source/WebCore/platform/mediastream/mac/ScreenCaptureKitSharingSessionManager.mm:
(WebCore::ScreenCaptureKitSharingSessionManager::cancelPendingSessionForDevice):
(WebCore::ScreenCaptureKitSharingSessionManager::contentFilterFromCaptureDevice):
(WebCore::ScreenCaptureKitSharingSessionManager::createSessionSourceForDevice):

Canonical link: https://commits.webkit.org/274943@main