REGRESSION(273406@main): Twitter/X inline video playback spins endlessly in Safari https://bugs.webkit.org/show_bug.cgi?id=269684 rdar://123190795
Reviewed by NOBODY (OOPS!).

In 273406@main, an incorrect assumption was made: that the state of MediaPlayer::supportsAcceleratedRendering
was the only factor that would influence the value of MediaPlayer::renderingCanBeAccelerated().
Previously, when the MediaPlayer's readyState changed, the AcceleratedRenderingStateChanged notification
to the GPUProcess would have been sent which would have triggered the decoding to start.
This notification was removed following 273406@main.

This change is a partial revert, to re-add the logic so that we check once again that the value of
renderingCanBeAccelerated() has changed and notify the GPUP if that is the case.

The original bug is still corrected, which would have been re-introduced with a full revert.

Manually tested, confirmed that the incorrect behaviour has been corrected and that the video
playback starts.
I haven't found a way to reproduce the bug with a test at this stage.

* Source/WebKit/WebProcess/GPU/media/MediaPlayerPrivateRemote.cpp:
(WebKit::MediaPlayerPrivateRemote::readyStateChanged):
(WebKit::MediaPlayerPrivateRemote::acceleratedRenderingStateChanged):
(WebKit::MediaPlayerPrivateRemote::checkAcceleratedRenderingState):
(WebKit::MediaPlayerPrivateRemote::updateConfiguration):
(WebKit::MediaPlayerPrivateRemote::setVideoFullscreenLayer):
* Source/WebKit/WebProcess/GPU/media/MediaPlayerPrivateRemote.h: