REGRESSION(274288@main): Null pointer dereference in `WebPageProxy::activityStateDidChange()` https://bugs.webkit.org/show_bug.cgi?id=269754 rdar://123032571
Reviewed by Chris Dumez.

`m_pageClient` is protected in `decidePolicyForNavigationAction()`, but not passed to the created
lambdas. When `Transaction` is destructed, `activityStateDidChange()` tries to create a strong reference
to `m_pageClient`, but it may already be destroyed. We should pass `protectedPageClient` into the
lambdas where `Transaction` is used.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::decidePolicyForNavigationAction):

Canonical link: https://commits.webkit.org/275019@main