[WGSL] Global sorting tries to print the name of a const_assert https://bugs.webkit.org/show_bug.cgi?id=269787 rdar://123238821
Reviewed by Mike Wyrzykowski.

A const_assert can never be part of a dependency cycle, but it can be visited
while the cycle is being computed. In that case, a debug logging was trying to
print the name of the const_assert, which resulted in an assertion failure.

* Source/WebGPU/WGSL/GlobalSorting.cpp:
(WGSL::nameForDeclaration):
(WGSL::Graph::addEdge):
(WGSL::reorder):

Canonical link: https://commits.webkit.org/275097@main