REGRESSION (Safari 17.2): Does not respect mousemove events in an iframe when the mouse is clicked from outside of an iframe https://bugs.webkit.org/show_bug.cgi?id=269886 rdar://120540148
Reviewed by Wenson Hsieh.

In 269246@main, we introduced a behavior change where initiating a drag
gesture from a cancelled mousedown event causes all mousemove events to
be routed to the originating frame till the drag gesture terminates with
a corresponding mouseup event. This was a web compatibility regression
since users expect to receive mouse events in the inner frame when we
are dragging the pointer over it.

Instead, this patch limits this mouse event target capturing behavior to
subframes only, and not the outer frame. This patch makes mouse event
targets independent of mousedown event cancellation when the pointer is
dragged out of a subframe. This change in behavior aligns us with other
major browser vendors' behavior.

Finally, we update cancel-mousedown-and-drag-to-frame.html since the new
behavior is essentially the opposite of what the test had been
expecting.

* LayoutTests/fast/events/cancel-mousedown-and-drag-from-frame-to-other-frame.html:
* LayoutTests/fast/events/cancel-mousedown-and-drag-to-frame-expected.txt:
* LayoutTests/fast/events/cancel-mousedown-and-drag-to-frame.html:
* LayoutTests/fast/events/resources/mouse-drag-from-frame-target-subframe.html:
* Source/WebCore/page/EventHandler.cpp:
(WebCore::EventHandler::handleMousePressEvent):

Canonical link: https://commits.webkit.org/275157@main